frontend-cloudrun/requirements.txt
Flask==2.0.2
psycopg2==2.9.3
Werkzeug==2.0.2
google-cloud-storage==1.44.0
google-cloud-secret-manager==2.8.0
Pillow==9.0.0
gunicorn==20.1.0